SDI-543 Merged tests into single test since there is only one actual resource now
